extract GGUF compatibility diagnostics into compat-core crate
