+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
   "crates/bitnet-bdd-grid",
   "crates/bitnet-bdd-grid-core",
   "crates/bitnet-compat",
+  "crates/bitnet-compat-core", # SRP: shared GGUF compatibility diagnostics core
   "crates/bitnet-ggml-ffi",     # GGML FFI for IQ2_S support
   "crates/bitnet-runtime-feature-flags",
   "crates/bitnet-runtime-feature-flags-core",
