+++ crates/bitnet-compat-core/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "bitnet-compat-core"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors = ["BitNet Contributors"]
+description = "SRP core GGUF compatibility diagnostics"
+license = "MIT OR Apache-2.0"
+
+[dependencies]
+anyhow.workspace = true
+bitnet-models = { path = "../bitnet-models", version = "0.2.1-dev" }
+tracing.workspace = true
+
+[dev-dependencies]
+tempfile.workspace = true
