+++ crates/bitnet-compat/src/gguf_fixer.rs
@@ -14,93 +14,7 @@ impl GgufCompatibilityFixer {
     /// Diagnose compatibility issues in a GGUF file
     pub fn diagnose<P: AsRef<Path>>(path: P) -> Result<Vec<String>> {
         let data = fs::read(path.as_ref())?;
-        let reader = GgufReader::new(&data)?;
-        let mut issues = Vec::new();
-
-        // Check tokenizer model
-        let tokenizer_model = reader
-            .get_string_metadata("tokenizer.ggml.model")
-            .unwrap_or_else(|| "unknown".to_string());
-
-        // Check for missing pre-tokenizer (critical for llama.cpp)
-        if tokenizer_model == "gpt2" || tokenizer_model == "llama3" {
-            let has_pre = reader.get_string_metadata("tokenizer.ggml.pre").is_some();
-            if !has_pre {
-                issues.push(format!(
-                    "Missing tokenizer.ggml.pre for {} tokenizer - llama.cpp will fail",
-                    tokenizer_model
-                ));
-            }
-        }
-
-        // Check for add_space_prefix
-        if tokenizer_model == "gpt2" {
-            let has_space_prefix =
-                reader.get_bool_metadata("tokenizer.ggml.add_space_prefix").is_some();
-            if !has_space_prefix {
-                issues.push("Missing tokenizer.ggml.add_space_prefix for GPT-2".to_string());
-            }
-        }
-
-        // Check special token IDs
-        let bos_id = reader.get_u32_metadata("tokenizer.ggml.bos_token_id");
-        let eos_id = reader.get_u32_metadata("tokenizer.ggml.eos_token_id");
-
-        if bos_id.is_none() {
-            issues.push("Missing BOS token ID".to_string());
-        }
-        if eos_id.is_none() {
-            issues.push("Missing EOS token ID".to_string());
-        }
-
-        // Check for vocabulary
-        if reader.get_u32_metadata("tokenizer.ggml.vocab_size").is_none() {
-            issues.push("No vocabulary found".to_string());
-        }
-
-        // BitNet-specific checks for GPT-2 tokenizer
-        if tokenizer_model == "gpt2" {
-            // Check tokens array length
-            if let Some(tokens) = reader.get_string_array_metadata("tokenizer.ggml.tokens") {
-                let vocab_len = tokens.len();
-                if vocab_len < 50_000 {
-                    issues.push(format!(
-                        "Vocabulary too small: {} tokens (expected >= 50,000 for GPT-2 family)",
-                        vocab_len
-                    ));
-                }
-            } else {
-                issues.push(
-                    "Missing tokenizer.ggml.tokens array - cannot verify vocabulary".to_string(),
-                );
-            }
-
-            // Check merges array exists (try both keys)
-            let has_merges = reader.get_string_array_metadata("tokenizer.ggml.merges").is_some()
-                || reader.get_string_array_metadata("tokenizer.ggml.bpe_merges").is_some();
-
-            if !has_merges {
-                issues.push(
-                    "Missing BPE merges (tokenizer.ggml.merges or tokenizer.ggml.bpe_merges)"
-                        .to_string(),
-                );
-            } else {
-                // Log merges count for diagnostics
-                let merges_count = reader
-                    .get_string_array_metadata("tokenizer.ggml.merges")
-                    .or_else(|| reader.get_string_array_metadata("tokenizer.ggml.bpe_merges"))
-                    .map(|m| m.len())
-                    .unwrap_or(0);
-
-                if merges_count == 0 {
-                    issues.push("BPE merges array is empty".to_string());
-                } else {
-                    info!("GPT-2 tokenizer has {} BPE merges", merges_count);
-                }
-            }
-        }
-
-        Ok(issues)
+        bitnet_compat_core::diagnose_data(&data)
     }
 
     /// Probe tokenization with a known phrase to verify first-token correctness
